Setup to install & test the package & tests
The installation is moved to a Makefile & this makes the installation
and testing easier and more elaborate.
The README is adjusted accordingly.

New features are test & coverage, which shows the % of lines that are
tested.

Tests check the conversion to namedtuple based on received responses
from the client.
If the client needs to be mocked, the client first gets configured
by the youtube_service.json response and then this mocked client
returns a required json.